Restructure the way config works.
I went to add a "vscode" key to the config to represent whether
the user's preferred editor is vscode, in which case we can
open it etc.

However, I quickly realised how horrible this process is and that
doesn't bode well for maintainability. So I've restructured
how config works, it is now a dataclass and the dict representation
of it as returned by pyyaml is now a typed dict which acts as sort of
a schema.

Hopefully, this is an improvement.