Add auto-detection of the conda envs folder.
In order to configure VSCode workspace with the correct python path
we need to know which conda (miniconda/anaconda) the user has installed.

Initially I thought about parsing the return from `conda info` but
I felt that would be unreliable and it's a lot of string
manipulation.

So I settled on a Naive check whether or not the appropriate
$HOME/miniconda3 (or anaconda3) directories existed.