Fix bug in config show on missing file.
There was a bug in config show and config set where
if the config file didn't exist, the `config.get()` method
would throw a `FileNotFoundError` which was not adequately
handled by the CLI causing an ugly raise to the user.

This commit fixes this bug and adds tests for it.