Fix conda env create if exists bug.
There was a bug where if you pytoil checkout a project
that already had an existing conda environment on the system,
it would raise a nasty error.

This exception is now handled and pytoil will now simply use the
existing environment as if it just created it.

This commit fixes this bug and adds tests for it.