Fix init ctrl+c bug.
This commit fixes a small bug with `pytoil init` whereby if the user
hit `ctrl+c` during the interactive config session, it would leave
the config file intact but with the default config parameters of
"UNSET" for `username` and `token`.

This in turn is checked by the `Config.validate()` method in later
functions but this simply raises an `InvalidConfigError` and was
never intended to be displayed to the user. This meant that
the user would get an ugly error and traceback which didn't necessarily
point to the root problem.

pytoil init now handles this by catching the `KeyboardInterrupt` and
deleting any config file before gracefully exiting the program.

This means that when the user runs pytoil init again, it works as if
the first call was never aborted.