Move all project subcommands into main.
From initial testing I decided that typing
`pytoil project checkout` etc was too clunky so I've moved
all of the commands that were inside `project` to the main
cli module now.

So now you can type `pytoil checkout`.

This will require a new version as it is a breaking change.