Remove CLI Tests.
I've decided to remove all the unit tests for the CLI. I found that
due to the nature of testing the CLI itself, the tests were far too
coupled to the implementation and it was difficult if not
impossible to test general behaviour because the CLI does things
like open vscode, clone git repos etc. so the majority of the tests
were full of mocks and `assert called` statements which means that
whenever I add functionality I have about an hour or so of tweaking
mocks. Not Ideal!

The CLI simply stitches together method calls that are themselves
well tested so it seemed redundant to me.